# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Cultour Project",
        "title": "Cultour API",
        "contact": {
            "name": "Cultour Development Team"
        },
        "version": "1.0"
    },
    "host": "localhost:8181",
    "paths": {
...
        "/badges": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a list of badges with pagination and filtering",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "badges"
                ],
                "summary": "List badges",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token (without 'Bearer ' prefix)",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of badges to retrieve",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of badges to skip",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "\"created_at\"",
                        "description": "Field to sort by",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "\"desc\"",
                        "description": "Sort order (asc/desc)",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Badges retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_achievement_models.Badge"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Failed to list badges",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add a new badge to the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "badges"
                ],
                "summary": "Create a new badge",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token (without 'Bearer ' prefix)",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "Badge Information",
                        "name": "badge",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_achievement_models.BadgeCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Badge created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_achievement_models.Badge"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid badge creation details",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/badges/count": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve the total number of badges in the system",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "badges"
                ],
                "summary": "Count badges",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token (without 'Bearer ' prefix)",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Badge count retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Failed to count badges",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/badges/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a badge by its unique identifier",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "badges"
                ],
                "summary": "Get a specific badge",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token (without 'Bearer ' prefix)",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Badge ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Badge retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_achievement_models.Badge"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Badge ID is required",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Badge not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update an existing badge by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "badges"
                ],
                "summary": "Update a badge",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token (without 'Bearer ' prefix)",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Badge ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Badge Update Information",
                        "name": "badge",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_achievement_models.BadgeCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Badge updated successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_achievement_models.Badge"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid badge update details or missing ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to update badge",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Remove a badge from the system by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "badges"
                ],
                "summary": "Delete a badge",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token (without 'Bearer ' prefix)",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Badge ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Badge ID is required",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to delete badge",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_holycann_cultour-backend_internal_achievement_models.Badge": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "icon_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_holycann_cultour-backend_internal_achievement_models.BadgeCreate": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "icon_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_holycann_cultour-backend_internal_response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Actual response data"
                },
                "error": {
                    "description": "Detailed error information (only populated for error responses)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.ErrorDetails"
                        }
                    ]
                },
                "message": {
                    "description": "Human-readable message",
                    "type": "string"
                },
                "pagination": {
                    "description": "Pagination information (optional)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.Pagination"
                        }
                    ]
                },
                "request_id": {
                    "description": "Unique request identifier for tracing",
                    "type": "string"
                },
                "status": {
                    "description": "Status of the response (success/error)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.ResponseStatus"
                        }
                    ]
                }
            }
        },
        "github_com_holycann_cultour-backend_internal_response.ErrorDetails": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Machine-readable error code",
                    "type": "string"
                },
                "details": {
                    "description": "Detailed error description",
                    "type": "string"
                }
            }
        },
        "github_com_holycann_cultour-backend_internal_response.Pagination": {
            "type": "object",
            "properties": {
                "has_next_page": {
                    "type": "boolean"
                },
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_holycann_cultour-backend_internal_response.ResponseStatus": {
            "type": "string",
            "enum": [
                "success",
                "error"
            ],
            "x-enum-varnames": [
                "StatusSuccess",
                "StatusError"
            ]
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Insert Your JWT Token. Do NOT include \"Bearer \" prefix. Example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}